Test classes refactored and bug fixed.
Hello test methods are separated to classes by endpoint.
Fixed the bug in Delete test to delete the created greeting.
