+++ src/test/java/com/labforward/api/hello/controller/HelloControllerPutTest.java
@@ -1,4 +1,4 @@
-package com.labforward.api.hello;
+package com.labforward.api.hello.controller;
 
 import static org.hamcrest.Matchers.contains;
 import static org.hamcrest.Matchers.hasSize;
@@ -24,72 +24,7 @@ import com.labforward.api.core.validation.EntityValidator;
 import com.labforward.api.hello.domain.Greeting;
 import com.labforward.api.hello.service.HelloWorldService;
 
-@AutoConfigureMockMvc
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class HelloControllerTest extends MVCIntegrationTest {
-	
-	private static final String EMPTY_MESSAGE = "";
-	private static final String HELLO_LUKE = "Hello Luke";
-	private static final String DUMMY_ID = "123";
-
-	@Test
-	public void getHelloIsOKAndReturnsValidJSON() throws Exception {
-		mockMvc.perform(get("/hello"))
-		       .andExpect(status().isOk())
-		       .andExpect(jsonPath("$.id", is(HelloWorldService.DEFAULT_ID)))
-		       .andExpect(jsonPath("$.message", is(HelloWorldService.DEFAULT_MESSAGE)));
-	}
-
-	@Test
-	public void postHelloReturnsUnprocessableEntityWhenMessageMissing() throws Exception {
-		String body = "{}";
-		mockMvc.perform(post("/hello").content(body)
-		                              .contentType(MediaType.APPLICATION_JSON))
-		       .andExpect(status().isUnprocessableEntity())
-		       .andExpect(jsonPath("$.validationErrors", hasSize(1)))
-		       .andExpect(jsonPath("$.validationErrors[*].field", contains("message")));
-	}
-
-	@Test
-	public void postHelloReturnsUnprocessableEntityWhenUnexpectedAttributeProvided() throws Exception {
-		String body = "{ \"tacos\":\"value\" }";
-		mockMvc.perform(post("/hello").content(body).contentType(MediaType.APPLICATION_JSON))
-		       .andExpect(status().isUnprocessableEntity());
-		// TODO(yigitalp): Does not append the MESSAGE_UNRECOGNIZED_PROPERTY. To be fixed.
-		//.andExpect(jsonPath("$.message", containsString(GlobalControllerAdvice.MESSAGE_UNRECOGNIZED_PROPERTY)));
-			
-	}
-
-	@Test
-	public void postHelloReturnsUnprocessableEntityWhenMessageEmptyString() throws Exception {
-		Greeting emptyMessage = new Greeting(EMPTY_MESSAGE);
-		final String body = getGreetingBody(emptyMessage);
-
-		mockMvc.perform(post("/hello").content(body)
-		                              .contentType(MediaType.APPLICATION_JSON))
-		       .andExpect(status().isUnprocessableEntity())
-		       .andExpect(jsonPath("$.validationErrors", hasSize(1)))
-		       .andExpect(jsonPath("$.validationErrors[*].field", contains("message")));
-	}
-	
-	@Test
-	public void postHelloReturnsBadRequestWhenBrokenJsonInputProvided() throws Exception {
-		String body = "{ \"tacos\":\"value\" ";
-		mockMvc.perform(post("/hello").content(body).contentType(MediaType.APPLICATION_JSON))
-		       .andExpect(status().isBadRequest());
-	}
-
-	@Test
-	public void postHelloIsCreatedWhenRequiredGreetingProvided() throws Exception {
-		Greeting hello = new Greeting(HELLO_LUKE);
-		final String body = getGreetingBody(hello);
-
-		mockMvc.perform(post("/hello").contentType(MediaType.APPLICATION_JSON)
-		                              .content(body))
-		       .andExpect(status().isCreated())
-		       .andExpect(jsonPath("$.message", is(hello.getMessage())));
-	}
+public class HelloControllerPutTest extends HelloControllerTest {
 	
 	@Test
 	public void putHelloReturnsUnprocessableEntityWhenIdMissing() throws Exception {
@@ -167,40 +102,6 @@ public class HelloControllerTest extends MVCIntegrationTest {
 		                              .content(body))
 		       .andExpect(status().isOk())
 		       .andExpect(jsonPath("$.message", is(hello.getMessage())));
-	}
-	
-	@Test
-	public void deleteHelloReturnsMethodNotAllowedWhenIdIsNotProvidedAsQueryParameter() throws Exception {
-		mockMvc.perform(delete("/hello").contentType(MediaType.APPLICATION_JSON))
-		       .andExpect(status().isMethodNotAllowed());
-	}
-
-	@Test
-	public void deleteHelloIsOKWhenRequiredGreetingProvided() throws Exception {
-		Greeting hello = new Greeting(HELLO_LUKE);
-		final String body = getGreetingBody(hello);
-		mockMvc.perform(post("/hello").contentType(MediaType.APPLICATION_JSON)
-		                              .content(body));
-
-		mockMvc.perform(delete("/hello/" + DUMMY_ID).contentType(MediaType.APPLICATION_JSON)
-													.content(body))
-			.andExpect(status().isOk());
-	}	
-	
-	@Test
-	public void deleteHelloIsNotFoundWhenGreetingDoesNotExist() throws Exception {
-		mockMvc.perform(delete("/hello/" + DUMMY_ID).contentType(MediaType.APPLICATION_JSON))
-			.andExpect(status().isNotFound());
-	}		
-
-	private String getGreetingBody(Greeting greeting) throws JSONException {
-		JSONObject json = new JSONObject().put("message", greeting.getMessage());
-
-		if (greeting.getId() != null) {
-			json.put("id", greeting.getId());
-		}
-
-		return json.toString();
 	}
 
 }
