+++ src/test/java/com/labforward/api/hello/controller/HelloControllerTest.java
@@ -1,0 +1,49 @@
+package com.labforward.api.hello.controller;
+
+import static org.hamcrest.Matchers.contains;
+import static org.hamcrest.Matchers.hasSize;
+import static org.hamcrest.Matchers.is;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.put;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+
+import org.json.JSONException;
+import org.json.JSONObject;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.http.MediaType;
+import org.springframework.test.context.junit4.SpringRunner;
+import org.springframework.test.web.servlet.MvcResult;
+import org.springframework.test.web.servlet.ResultActions;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.labforward.api.common.MVCIntegrationTest;
+import com.labforward.api.core.validation.EntityValidator;
+import com.labforward.api.hello.domain.Greeting;
+import com.labforward.api.hello.service.HelloWorldService;
+
+@AutoConfigureMockMvc
+@RunWith(SpringRunner.class)
+@SpringBootTest
+public abstract class HelloControllerTest extends MVCIntegrationTest {
+	
+	protected static final String EMPTY_MESSAGE = "";
+	protected static final String HELLO_LUKE = "Hello Luke";
+	protected static final String DUMMY_ID = "123";
+
+	protected String getGreetingBody(Greeting greeting) throws JSONException {
+		JSONObject json = new JSONObject().put("message", greeting.getMessage());
+
+		if (greeting.getId() != null) {
+			json.put("id", greeting.getId());
+		}
+
+		return json.toString();
+	}
+}
